+++ js/packages/openinference-instrumentation-langchain/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @arizeai/openinference-instrumentation-langchain
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [9c44b14]
+  - @arizeai/openinference-semantic-conventions@0.7.0
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [b66bf54]
+- Updated dependencies [fe69250]
+- Updated dependencies [60ade67]
+  - @arizeai/openinference-semantic-conventions@0.6.0
+
 ## 0.0.6
 
 ### Patch Changes